+++ src/publication/3D.ts
@@ -4,11 +4,12 @@ import { PublicationMainFocus } from './PublicationMainFocus.js';
 import { PublicationSchemaId } from './PublicationSchemaId.js';
 import {
   AnyMediaSchema,
+  mainContentFocus,
   metadataDetailsWith,
   MetadataLicenseTypeSchema,
   publicationWith,
 } from './common';
-import { notEmptyString, uriSchema } from '../primitives.js';
+import { nonEmptyStringSchema, uriSchema } from '../primitives.js';
 
 export enum ThreeDFormat {
   gLTF = 'gLTF/GLB',
@@ -22,7 +23,7 @@ export enum ThreeDFormat {
  */
 export const ThreeDAssetSchema = z.object({
   uri: uriSchema('The 3D asset url or zip'),
-  zipPath: notEmptyString()
+  zipPath: nonEmptyStringSchema()
     .optional()
     .describe('path in extracted zip. Relative. 3D start point, must be 3D file type'),
   playerUrl: uriSchema('Link to web based 3D player'),
@@ -35,7 +36,7 @@ export const ThreeDMetadataSchema = publicationWith({
   $schema: z.literal(PublicationSchemaId.THREE_D_LATEST),
 
   lens: metadataDetailsWith({
-    mainContentFocus: z.literal(PublicationMainFocus.THREE_D),
+    mainContentFocus: mainContentFocus(PublicationMainFocus.THREE_D),
 
     assets: ThreeDAssetSchema.array().min(1).describe('The 3D items for the publication'),