+++ src/publication/EmbedSchema.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 import { PublicationMainFocus } from './PublicationMainFocus.js';
 import { PublicationSchemaId } from './PublicationSchemaId.js';
 import { AnyMediaSchema, mainContentFocus, metadataDetailsWith, publicationWith } from './common';
-import { uriSchema } from '../primitives.js';
+import { encryptableUriSchema } from '../primitives.js';
 
 /**
  * @internal
@@ -14,7 +14,7 @@ export const EmbedSchema = publicationWith({
   lens: metadataDetailsWith({
     mainContentFocus: mainContentFocus(PublicationMainFocus.EMBED),
 
-    embed: uriSchema('The embed URL.'),
+    embed: encryptableUriSchema('The embed URL.'),
 
     attachments: AnyMediaSchema.array()
       .min(1)