+++ src/publication/LivestreamSchema.ts
@@ -3,7 +3,11 @@ import { z } from 'zod';
 import { PublicationMainFocus } from './PublicationMainFocus.js';
 import { PublicationSchemaId } from './PublicationSchemaId.js';
 import { AnyMediaSchema, mainContentFocus, metadataDetailsWith, publicationWith } from './common';
-import { datetimeSchema, notEmptyString, uriSchema } from '../primitives.js';
+import {
+  encryptableDatetimeSchema,
+  encryptableUriSchema,
+  nonEmptyStringSchema,
+} from '../primitives.js';
 
 /**
  * @internal
@@ -13,25 +17,27 @@ export const LivestreamSchema = publicationWith({
   lens: metadataDetailsWith({
     mainContentFocus: mainContentFocus(PublicationMainFocus.LIVESTREAM),
 
-    title: notEmptyString().optional().describe('The livestream title.'),
+    title: nonEmptyStringSchema().optional().describe('The livestream title.'),
 
-    startsAt: datetimeSchema('The stream start time (ISO 8601 `YYYY-MM-DDTHH:mm:ss.sssZ`).'),
+    startsAt: encryptableDatetimeSchema(
+      'The stream start time (ISO 8601 `YYYY-MM-DDTHH:mm:ss.sssZ`).',
+    ),
 
-    endsAt: datetimeSchema(
+    endsAt: encryptableDatetimeSchema(
       'The optional stream end time (ISO 8601 `YYYY-MM-DDTHH:mm:ss.sssZ`)',
     ).optional(),
 
-    playbackUrl: uriSchema(
+    playbackUrl: encryptableUriSchema(
       'Some livestream platforms have the playback url as a separate url. ' +
         'If not your case make sure `liveUrl` and `playbackUrl` are the same.',
     ),
 
-    liveUrl: uriSchema(
+    liveUrl: encryptableUriSchema(
       'Some livestream platforms have the live url as a separate url. ' +
         'If not your case make sure `liveUrl` and `playbackUrl` are the same.',
     ),
 
-    checkLiveAPI: uriSchema(
+    checkLiveAPI: encryptableUriSchema(
       'The data cannot be changed so you can put in an API endpoint to know if it is still live or not for clients to be able to check.',
     ).optional(),