+++ src/publication/VideoSchema.ts
@@ -9,7 +9,7 @@ import {
   metadataDetailsWith,
   publicationWith,
 } from './common';
-import { notEmptyString } from '../primitives.js';
+import { nonEmptyStringSchema } from '../primitives.js';
 
 /**
  * @internal
@@ -24,7 +24,7 @@ export const VideoSchema = publicationWith({
 
     video: MediaVideoSchema,
 
-    title: notEmptyString('The optional video title.').optional(),
+    title: nonEmptyStringSchema('The optional video title.').optional(),
 
     attachments: AnyMediaSchema.array()
       .min(1)